mangd/requirements.txt
Authlib==0.15.5
Flask==2.0.2
Flask_Mail==0.9.1
Flask_Session==0.4.0
Flask_SQLAlchemy==2.5.1
google_api_python_client==2.33.0
google_auth_oauthlib==0.4.6
itsdangerous==2.0.1
protobuf==3.19.1
Werkzeug==2.0.2
